create image-handler and bring image-uploader to previeous version
